Moved FeatureTest, FeatureTestReplace, FeatureTestAddition into tests.base
